UI is in place, renamed several classes and better stablished metadata classes. Updated everything to work with the new system.

+++ Master/NucleusGaming/Globals.cs
@@ -7,7 +7,7 @@ namespace Nucleus.Gaming
 {
     public static class Globals
     {
-        public const decimal Version = 9.0M;
+        public const uint Version = 9;
         public const bool Alpha = true;
     }
 }
